Force scrollbar inside xxx2Many table
On Chrome based browsers the min-content of the table inside a xxx2Many widget
is the minimal size to display all the columns.
So when this min-content size is bigger than the width of the tab, it is the
all form that get a scrollbar instead of only the tree view element.
By adding a display grid to the div containing the treeview, the min-content
is not more used by the grid-template-columns of the form-container.

Closes #13050