Fetch URL icons asynchronously
If the URL is not yet in cache, an empty image at the proper size is return,
then when the callback is triggered, the URL should be in cache. So we force a
redraw of the treeview which should set the proper pixbuf.
The IconFactory uses a pool of 5 workers like the browsers.

Closes #13119