Set posting employee only when using button
As the posted by employee field can not be modified once the invoice is
posted, the employee must be set only when it is an action of the user.
On processing the invoice, it is the internal posting method that must be used
so no employee is set nor any other warning checks.

Closes #13124
(grafted from 361b4624d10c7173ce439329d72e37d7fa02d00b)

--HG--
branch : 7.0