Do not try to round or convert list price of project revenue
By using list_price_used we are sure to have a list price in the company
currency so there is not need for conversion.
Also list price of a product can be empty so we should not try to round it.

Closes #13110
(grafted from 710927560bfe722fc02c0d22bbdf2c1615f787a7)

--HG--
branch : 6.8