Use file from the Binary field to display document
On Windows the file used to create the Evince document must be kept after the
initialization.
Also reusing the file cache mechanism of the Binary field, prevents to create
temporary files on each call of display.
Evince document needs to have the proper extension to be sure to load the
proper backend. So the cache file must be retrieved with the requested
extension.
And finally we use Document.factory_get_document_full so if leading with the
filename extension does not work, it will try to guess from the data.

Closes #13307
(grafted from 34deeaf3c0215f793c3e673de3ee2e04971ebee4)

--HG--
branch : 7.2