Reuse simple COALESCE expression to filter moves after stock period
The expression must be the same as the one in the indexes such that the
indexes can be used to directly filter out all the moves before the period.
This restores the original expression that was changed by e4431b79e409 for the
generic move_date. But by design there can not be any assigned move planned
before the closed period.