Do not allow to mutate dict value
A copy of the value is always returned so if it is modified it does not modify
the internal value of the record.
So to update Dict value, a new dictionary must be set on the field.

Closes #13374
(grafted from 82b0280d4a08c48b2ab85b4ababe2031aee9a7a2)

--HG--
branch : 7.2