Ensure effective date is before or after origin or outcome stock moves
When the source of the move is the same as the destination of its origin or
outcomes, the stock move can not be done before those moves.
When the destination of the move is the same as the source of the its origin
or outcomes, the stock move can not be done after those moves.

Closes #12851