Ensure current record is None after removing record from the group
The removal can trigger __select_changed on the list view which may set the
first record of the selection as the current record which may be the deleted
record.
This prevents the display triggered by record_modified to set the current
record in the selection as it has been removed from the group.

Closes #13712
(grafted from ed532372fe23eff7fe3e7c94b145bd87c5cf741d)

--HG--
branch : 6.0