Include the depends of the getter function when depending on a Function field
Since c47b3642f1a6, the getter of a Function field may be called if there is no
value set for it on the instance.
In the on_change context, this may set the missing fields to None if they were
not defined in the depends of the method called. In addition to compute a wrong
value, those fields are then returned as modified.

Closes #13735
(grafted from 1b815539165a1212fc5c415c8920d8f5f0c55236)

--HG--
branch : 6.0