Add deletion on FileStore
We're identifying each file in the filestore with a UUID4 in order to avoid
collisions. It has the drawback that we will store twice the same file but on
the other hand we can now remove each file from the store as there is no more
collision.

In order to identify files named previously we're using textual representation
of the UUID with the hyphens. Previous IDs do not have hyphens.