Use company from the record when copying budget
Since ba6b8ce0e692 only the fields displayed on the form state are set but the
copy wizard for analytic budget does not have any field depending on the
company so no default value is set for company.
Indeed we should not rely on the state view to retrieve the currency from the company.

Closes #13824