Compute digits of float value before applying factor
When applying the factor, we introduce new rounding issue which falsify the
computation of digits.
So it is better to compute the digits using the initial value (coming from the
string parsing) and update the digits based on the order of the factor.

Closes #13964
(grafted from 3606a9f176d0312d9645d05902b74192391538e4)

--HG--
branch : 7.4