Process on a copy of imported data
The process_lines of ModelStorage.import_data pop rows when importing One2Many.
But if the transaction has to be restarted for example if a lock is needed, the
original data from the request has been altered for the second call losing some
or all the data to be imported.

Closes #13984
(grafted from 2cb96c1427d7c0edcb1c9b5b410e220db3f8958e)

--HG--
branch : 7.0