Do not fail to reload modified view without a current record
It may happen that a tree view is considered as modified even if there is no
current record. In this case we should not try to compare the id of the current
record before and after reload as it is not defined.

Closes #13998
(grafted from 331d52d447992f2bc923df8cd8bf88690460f637)

--HG--
branch : 7.4