Set email content with maintype when needed and correctly set Content-Language header
When the content is not a string, EmailMessage.set_content requires a maintype
argument but not when it is a string. The maintype is then deduced using the
mimetypes module.
The Content-Language of content or alternate must be set using the headers
parameter instead of params which is for the Content-Type.
And finally the Content-Language must be set on the main message only if it
contains multiparts otherwise it is set twice.

Closes #14028
(grafted from 27bd4c93c85b3e64894ea97f4e433fb9fc3579ba)

--HG--
branch : 7.4