Round only the tax amount of the line when rounding per line
If we round the tax amount already added to the other line, the result may vary
depending on the order of the lines.
This is because round half even is used so the result may depend on the already
summed tax amounts.

Closes #14047
(grafted from e1e376f91bc1e111cdef39ddcbbfd92d201ea6c9)

--HG--
branch : 7.0