Set selected class before changing current record and do not rely on toggle
The class must be set before changing the current record because the property
selected_records relies on the classes and this method is called when changing
the current record to set the sensitivity of the buttons.
Also clicking a second time on the row must unselect the row which is not
possible to implement with the toggling of the class.

Closes #13980