Remove default lead time for product sold
By default it is better to assume that we do not know the default lead time of
the company which allows to select the first supplier no matter its lead time.
The sale already plans the stock move to today when the planned shipping date
is empty since a748de86c8d6.

Closes #14224