Set default value of src_plural to NULL when setting translations
Adding the new translations to the internal language should set the value of
src_plural to NULL as this is the default value we use when importing from the
PO files. Moreover it might happen that an empty string could be a valid value
for src_plural.

Closes #14289
(grafted from d3380ee74626e57f1e265034ce6689e8e3971ccd)

--HG--
branch : 7.6