Do not prevent to process sale line without product and quantity from Shopify
When a line is removed on Shopify, the quantity of the sale line is set to 0 in
Tryton because Shopify refund the line.
So such line without product must not prevent to process the sale as they will
not generate any stock move nor invoice line.

Closes #14375
(grafted from bbd807ca4b87dc0470bd5f3db69681e7371c6d13)

--HG--
branch : 7.4